+++ game/effects/src/lua/banner_effect_2.c
@@ -31,8 +31,8 @@ static void config(lua_State *L, struct banner * banner,struct rect * rect  ) {
 	char * txt  = (char *) luaL_checkstring (L, -6);
  	int start_idx  = (int) luaL_checknumber (L, -5);
  	int end_idx  = (int) luaL_checknumber (L, -4);
- 	int on_time  = (int) luaL_checknumber (L, -3);
- 	int off_time  = (int) luaL_checknumber (L, -2);
+ 	double on_time  = (double) luaL_checknumber (L, -3);
+ 	double off_time  = (double) luaL_checknumber (L, -2);
  	int font  = (int) luaL_checknumber (L, -1);
